+++ apps/webapp/components/settings/space/components/AddMoreMemberProfilesModal.tsx
@@ -58,7 +58,7 @@ export function AddMoreMemberProfilesModal({
 export function getProfileWidgetLogo(name: MemberProfileName) {
   switch (name) {
     case 'ZAO':
-      return '/images/logos/KFMEDIA_black.png';
+      return '/images/logos/charmverse_black.png';
     case 'collection':
       return '/images/template_icons/nft_ape_icon.svg';
     case 'ens':